# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Clotho API Gateway
  description: |
    Product REST contract for Clotho (`/api/v1/*`). Humans use the CLI; agents
    use MCP tools that call this same edge; the web app and `@clotho/sdk-js`
    share it. Errors use `{ "error": "..." }` with 4xx/5xx status codes.

    **Auth:** pluggable `AuthProvider` (ADR-0018). Default `bootstrap` accepts
    `Authorization: Bearer clotho_tok_…` and, when `CLOTHO_AUTH_REQUIRED=false`,
    falls back to a bootstrap actor. Managed profiles set
    `CLOTHO_AUTH_REQUIRED=true` and `CLOTHO_AUTH_PROVIDER=clerk` (Clerk session
    JWT / org API keys; Clotho `clotho_tok_…` still accepted — §11 #7).
    Mint Clotho tokens with `POST /api/v1/tokens` or `clotho auth token create`.
    Agents never use Clerk — MCP uses scoped `clotho_agt_…` tokens minted via
    human-only `/api/v1/agents/*` (ADR-0016). Secrets master key is server-side
    only (`CLOTHO_SECRETS_MASTER_KEY`).

    **Provider Fabric (Stage 17):** `GET /api/v1/providers?layer=compute|storage|network|auth`
    (omit `layer` for compute-only backward compatibility; `?all=true` for every
    layer). Storage/network are honest stubs until Stages 18–19.
  version: 0.1.0
  license:
    name: Apache-2.0
servers:
  - url: http://localhost:8080
    description: Local dev stack (`just dev`)
...
paths:
...
  /api/v1/providers:
    get:
      tags: [providers]
      operationId: listProviders
      summary: Provider Fabric registry (compute by default)
      description: |
        Without `layer`, returns the compute provider registry (Stage 12 shape).
        With `layer=compute|storage|network|auth`, returns fabric metadata for
        that layer. With `all=true`, returns every layer. Storage/network stubs
        report `configured: false` until Stages 18–19 (ADR-0019).
      parameters:
        - name: layer
          in: query
          required: false
          schema:
            type: string
            enum: [compute, storage, network, auth]
          description: Fabric layer filter
        - name: all
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: When true and layer omitted, include all fabric layers
      responses:
        "200":
          description: Providers
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ProviderList"
                  - $ref: "#/components/schemas/FabricProviderList"
        "400":
          $ref: "#/components/responses/Error"
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Human API token (`clotho_tok_…`)
  responses:
    Error:
      description: Error envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    ProviderCapabilities:
      type: object
      properties:
        one_shot_jobs:
          type: boolean
        persistent_workspaces:
          type: boolean
        snapshots:
          type: boolean
        templates:
          type: boolean
        regions:
          type: array
          items:
            type: string
        ssh:
          type: boolean
        desktop:
          type: boolean
        public_url:
          type: boolean
        file_api:
          type: boolean
        terminal_streaming:
          type: boolean
        cost_hints:
          type: string
    Provider:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        kind:
          type: string
        enabled:
          type: boolean
        configured:
          type: boolean
        configured_reason:
          type: string
        capabilities:
          type: array
          items:
            type: string
        capability_detail:
          $ref: "#/components/schemas/ProviderCapabilities"
        default_snapshot:
          type: string
        notes:
          type: string
    ProviderList:
      type: object
      properties:
        providers:
          type: array
          items:
            $ref: "#/components/schemas/Provider"
        default_provider_id:
          type: string
    FabricProvider:
      type: object
      description: Shared Provider Fabric metadata (ADR-0019); never includes secrets
      properties:
        id:
          type: string
        name:
          type: string
        layer:
          type: string
          enum: [compute, storage, network, auth]
        kind:
          type: string
        enabled:
          type: boolean
        configured:
          type: boolean
        configured_reason:
          type: string
        capabilities:
          type: array
          items:
            type: string
        notes:
          type: string
    FabricProviderList:
      type: object
      properties:
        providers:
          type: array
          items:
            $ref: "#/components/schemas/FabricProvider"
        default_provider_id:
          type: string
        layer:
          type: string
          nullable: true
...